requirements.txt
-e lib/ape-zksync
black
eth-ape
flake8
isort
pre-commit
